+++ apps/helpers.py
@@ -1,100 +1,15 @@
-import re
-import time
-import uuid
 from datetime import datetime
 from enum import Enum
 from typing import Optional
 
-from django.apps import apps
-from django.conf import settings
-from django.core import serializers
-from django.core.exceptions import ObjectDoesNotExist, ValidationError
-from django.core.validators import RegexValidator
+from django.core.exceptions import ObjectDoesNotExist
 from django.db import transaction
-from django.template import engines
 
-from projects.models import Flavor
 from studio.utils import get_logger
 
-from .models import (
-    Apps,
-    AppStatus,
-    BaseAppInstance,
-    JupyterInstance,
-    Subdomain,
-    VolumeInstance,
-)
-from .tasks import delete_resource, deploy_resource
+from .models import Apps, AppStatus, BaseAppInstance, Subdomain
 
 logger = get_logger(__name__)
-
-
-HELP_MESSAGE_MAP = {
-    "name": "Display name for the application. This is the name visible on the app catalogue if the app is public",
-    "description": "Summarize the application in a few words",
-    "subdomain": "Valid subdomain names have minimum length of 3 characters and may contain lower case letters a-z \
-        and numbers 0-9 and a hyphen '-'. The hyphen should not be at the start or end of the subdomain.",
-    "access": "Public apps will be displayed on the app catalogue and can be accessed by anyone that has the link to \
-        them. Project apps can only be accessed by project members. Private apps are only accessible by users that \
-        create the apps.",
-    "source_code_url": "Provide a URL where the full source code of your app can be accessed (for example, to a GitHub \
-        repository or a Figshare or Zenodo entry).",
-    "flavor": "Hardware allocation for your app. Only one option is available by default. If your app requires more \
-        hardware resources, get in touch with us (serve@scilifelab.se) with a request.",
-    "image": "Docker Image for the app uploaded to DockerHub or GitHub. Each version of your app should have a unique \
-        tag.",
-    "path": "Specify the path inside the container that you want to be persistent (path to database or similar). If \
-        you follow our guide to build the container, then please include the username in the path as well.",
-    "port": "Port that the docker container exposes and the application runs on. This should be an integer between \
-        3000-9999.",
-    "note_on_linkonly_privacy": "This option can be used only for a limited amount of time, for example while under \
-        development or during peer review.",
-}
-
-
-class SubdomainCandidateName:
-    """
-    A candidate subdomain name that may or may not be available or allowed.
-    """
-
-    __name = None
-
-    def __init__(self, name: str):
-        self.__name = name
-
-    def is_available(self) -> bool:
-        """Determines if the candidate name is available in Serve."""
-        if self.__name == "serve":
-            # Reserved words
-            return False
-        elif Subdomain.objects.filter(subdomain=self.__name).exists():
-            return False
-        else:
-            return True
-
-    def is_valid(self) -> bool:
-        """Determines if the candidate name is valid."""
-        try:
-            self.validate_subdomain()
-            return True
-        except ValidationError:
-            return False
-
-    def validate_subdomain(self):
-        """
-        Validates a subdomain text.
-        The RegexValidator will raise a ValidationError if the input does not match the regular expression.
-        It is up to the caller to handle the raised exception if desired.
-        """
-
-        # Check if the subdomain adheres to helm rules
-        regex_validator = RegexValidator(
-            regex=r"^(?!-)[a-z0-9-]{3,53}(?<!-)$",
-            message="Subdomain must be 3-53 characters long, contain only lowercase letters, digits, hyphens, "
-            "and cannot start or end with a hyphen",
-        )
-
-        regex_validator(self.__name)
 
 
 def can_access_app_instance(instance, user, project):
@@ -314,6 +229,7 @@ def create_instance_from_form(form, project, app_slug, app_id=None):
     Raises:
     - ValueError: If the form does not have a 'subdomain' or if the specified app cannot be found.
     """
+    from .tasks import deploy_resource
 
     subdomain_name = get_subdomain_name(form)
 
@@ -350,6 +266,8 @@ def get_or_create_status(instance, app_id):
 
 
 def handle_subdomain_change(instance, subdomain, subdomain_name):
+    from .tasks import delete_resource
+
     if instance.subdomain.subdomain != subdomain_name:
         # In this special case, we avoid async task.
         delete_resource(instance.serialize())
