+++ cypress/e2e/setup-scripts/seed_collections_user.py
@@ -7,7 +7,7 @@ from django.conf import settings
 from django.contrib.auth.models import User
 from django.db import transaction
 
-from apps.constants import APP_REGISTRY
+from apps.app_registry import APP_REGISTRY
 from apps.helpers import create_instance_from_form
 from apps.models import Apps
 from projects.models import Flavor, Project, ProjectTemplate
