+++ projects/views.py
@@ -22,7 +22,7 @@ from django.views import View
 from guardian.decorators import permission_required_or_403
 from guardian.shortcuts import assign_perm, remove_perm
 
-from apps.constants import APP_REGISTRY
+from apps.app_registry import APP_REGISTRY
 from apps.models import BaseAppInstance
 
 from .exceptions import ProjectCreationException
